share utils instead of duplication
