+++ src/cpp/benders/output/CMakeLists.txt
@@ -9,7 +9,6 @@
 add_library (output_core STATIC
 	${CMAKE_CURRENT_SOURCE_DIR}/VoidWriter.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/JsonWriter.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/Clock.cpp
         )
 
 get_target_property(JSON_INC_PATH jsoncpp_lib INTERFACE_INCLUDE_DIRECTORIES)
@@ -27,6 +26,7 @@ target_include_directories (output_core
 target_link_libraries (output_core
 	PUBLIC
 		jsoncpp_lib
+		helpers
 )
 
 add_library (${PROJECT_NAME}::output_core ALIAS output_core)
