--- src/cpp/benders/output/Clock.cpp
@@ -1,9 +1,0 @@
-#include "Clock.h"
-
-#include <ctime>
-#include <string>
-
-std::time_t Clock::getTime() {
-  std::time_t beginTime = std::time(nullptr);
-  return beginTime;
-}
