+++ src/cpp/benders/output/JsonWriter.cpp
@@ -4,19 +4,6 @@
 
 #include "config.h"
 
-namespace clock_utils {
-std::string timeToStr(const std::time_t &time_p) {
-  struct tm local_time;
-  localtime_platform(time_p, local_time);
-  // localtime_r(&time_p, &local_time); // Compliant
-  const char *FORMAT = "%d-%m-%Y %H:%M:%S";
-  char buffer_l[100];
-  strftime(buffer_l, sizeof(buffer_l), FORMAT, &local_time);
-  std::string strTime_l(buffer_l);
-
-  return strTime_l;
-}
-}  // namespace clock_utils
 namespace Output {
 JsonWriter::JsonWriter(std::shared_ptr<Clock> p_clock,
                        const std::filesystem::path &json_filename)
