--- src/cpp/benders/output/include/Clock.h
@@ -1,12 +1,0 @@
-
-#pragma once
-
-#include <ctime>
-
-class Clock {
- public:
-  Clock() = default;
-  virtual ~Clock() = default;
-
-  virtual std::time_t getTime();
-};
